requirements.txt
boto3 >= 1.9.160
argcomplete >= 1.8.1
auth0-python >= 3.13.0
